requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.5
django-cors-headers==3.2.1
djangorestframework==3.11.0
idna==2.9
pycryptodome==3.9.7
PySimpleGUI==4.18.2
pytz==2019.3
requests==2.23.0
sqlparse==0.3.1
urllib3==1.25.9
